Update double gyre base case for evaluation metric
The diagnostic table has been in the test case has been updated to
allow for the 'jet penetration' metric described in McCalpin 1995,
calculated as the first moment of kinetic energy with respect to
longitude. A notebook has been added demonstrating how this metric
should be calculated.

Some additional diagnostics are included at 5day, monthly, and
annual frequencies to help determine whether the model has come
into a pseudo-steady state and to evaluate the overall behavior
of the model.